+++ marketplace/src/main/java/com/boost/marketplace/ui/popup/customdomains/ConfirmedCustomDomainBottomSheet.kt
@@ -321,7 +321,7 @@ class ConfirmedCustomDomainBottomSheet(val activity: CustomDomainActivity) :
         return UserSessionManager(requireContext()).getAccessTokenAuth()?.barrierToken() ?: ""
     }
 
-    private fun getDiscountedPrice(price: Double, discountPercent: Int): Double {
+    private fun getDiscountedPrice(price: Double, discountPercent: Double): Double {
         return price - ((discountPercent / 100) * price)
     }
